requirements.txt
apache-airflow-client==2.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
requests==2.31.0
rich==13.7.1
secure-smtplib==0.1.1
six==1.16.0
tabulate==0.9.0
urllib3==2.2.1
